Wrapped service in index file
